d2fsam-web-app/src/main/webapp/app/entities/external-file-resource/delete/external-file-resource-delete-dialog.component.ts
import { Component } from '@angular/core';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { IExternalFileResource } from '../external-file-resource.model';
import { ExternalFileResourceService } from '../service/external-file-resource.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  templateUrl: './external-file-resource-delete-dialog.component.html',
})
export class ExternalFileResourceDeleteDialogComponent {
  externalFileResource?: IExternalFileResource;

  constructor(protected externalFileResourceService: ExternalFileResourceService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.externalFileResourceService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
